+++ includes/data-feeds.php
@@ -45,6 +45,12 @@
 // - Not Found Feed Error
 // - Format Data to XML
 // - Convert Array to XML
+// === RSS Feeds ===
+// - Show Posts Feed Conflict Fix
+// - Show Posts Feed filter
+// - Add Feed Item Show Node
+// - Add Feed Item Host Node
+// - Add Feed Item Producer Node
 // === Shift Conversions ===
 // - Convert Show Shift
 // - Convert Show Shifts
@@ -1337,6 +1343,147 @@ function radio_station_feed_not_found( $error ) {
 		),
 	);
 	return $error;
+}
+
+
+// -----------------
+// === RSS Feeds ===
+// -----------------
+
+// ----------------------------
+// Show Posts Feed Conflict Fix
+// ----------------------------
+// 2.5.5: added to allow filtering posts by related show
+add_filter( 'parse_query', 'radio_station_feed_filter_fix', 0 );
+function radio_station_feed_filter_fix( $query ) {
+
+	// --- override incorrect shows feed ---
+	if ( isset( $query->query['feed'] ) && ( 'feed' == $query->query['feed'] ) && ( 'show' == $query->query['post_type'] ) ) {
+		
+		if ( strstr( $_SERVER['REQUEST_URI'], '/shows/feed/' ) ) {
+			
+			// --- add host/producer nodes to RSS item output ---
+			add_action( 'rss2_item', 'radio_station_feed_item_node_hosts' );
+			add_action( 'rss2_item', 'radio_station_feed_item_node_producers' );
+			
+		} else {
+
+			// --- fix for post feed with show filter ---
+			$query->query['post_type'] = 'post';
+			$query->query_vars['post_type'] = 'post';
+			$query->query_vars['name'] = '';
+			$query->is_comment_feed = '';
+			$query->is_post_type_archive = '1';
+			unset( $query->query['name'] );
+
+			// --- add show node to RSS item output ---
+			add_action( 'rss2_item', 'radio_station_feed_item_node_shows' );
+
+		}
+		
+	} 
+	// print_r( $query );
+}
+
+// ----------------------
+// Show Posts Feed Filter
+// ----------------------
+// eg. /feed/?show=something
+// 2.5.5: added to allow filtering posts by related show
+add_filter( 'pre_get_posts', 'radio_station_feed_filter' );
+function radio_station_feed_filter( $query ) {
+
+	// --- check for shows query parameter on posts feed ---
+    if ( $query->is_feed && $query->is_main_query() && isset( $query->query['post_type'] ) && ( 'post' == $query->query['post_type'] ) && isset( $query->query_vars['show'] ) ) {
+		$show_id = $query->query_vars['show'];
+		$show = get_post( $show_id );
+		if ( !$show ) {
+			// --- get show by post slug ---
+			global $wpdb;
+			$q = "SELECT ID FROM " . $wpdb->prefix . "posts WHERE post_name = %s AND post_type = %s";
+			$q = $wpdb->prepare( $q, array( $show_id, RADIO_STATION_SHOW_SLUG ) );
+			$show_id = $wpdb->get_var( $q );
+		}
+		if ( $show_id ) {
+			$query->set( 'meta_query', array(
+				array(
+					'key'       => 'post_showblog_id',
+					'value'     => $show_id,
+					'compare'   => 'EQUALS'
+				)
+			) );
+		}
+    }
+
+    return $query;
+}
+
+// -----------------------
+// Add Feed Item Show Node
+// -----------------------
+// 2.5.5: added show node to post feed items
+function radio_station_feed_item_node_shows() {
+	global $post;
+	$show_id = get_post_meta( $post->ID, 'post_showblog_id', true );
+	if ( $show_id ) {
+		$show = get_post( $show_id );
+		echo '<show>' . esc_html( $show->post_title ) . '</show>' . PHP_EOL;
+		echo '<show-id>' . esc_html( $show_id ) . '</show-id>' . PHP_EOL;
+	}
+}
+
+// -----------------------
+// Add Feed Item Host Node
+// -----------------------
+// 2.5.5: add host node to show feed items
+function radio_station_feed_item_node_hosts() {
+	global $post;
+	$host_ids = get_post_meta( $post->ID, 'show_user_list', true );
+	$hosts = '';
+	$count = 0;
+	if ( $host_ids && is_array( $host_ids ) && ( count( $host_ids ) > 0 ) ) {
+		$host_count = count( $hosts_ids );
+		foreach ( $host_ids as $host ) {
+			$count++;
+			$user = get_user_by( 'ID', $host );
+			$hosts .= $user->display_name;
+			if ( ( ( 1 == $count ) && ( 2 == $host_count ) ) || ( ( $host_count > 2 ) && ( ( $host_count - 1 ) == $count ) ) ) {
+				$hosts .= ' ' . __( 'and', 'radio-station' ) . ' ';
+			} elseif ( ( $count < $host_count ) && ( $host_count > 2 ) ) {
+				$hosts .= ', ';
+			}
+		}
+	}
+	if ( '' != $hosts ) {
+		echo '<host>' . esc_html( $hosts ) . '</host>' . PHP_EOL;
+	}
+}
+
+// ---------------------------
+// Add Feed Item Producer Node
+// ---------------------------
+// 2.5.5: add producer node to show feed items
+function radio_station_feed_item_node_producers() {
+	global $post;
+	$producer_ids = get_post_meta( $post->ID, 'producer_user_id', true );
+	$producers = '';
+	$count = 0;
+	if ( $producer_ids && is_array( $producer_ids ) && ( count( $producer_ids ) > 0 ) ) {
+		$producer_count = count( $producer_ids );
+		foreach ( $producer_ids as $producer ) {
+			$count++;
+			$user = get_user_by( 'ID', $producer );
+			$producers .= $user->display_name;
+			if ( ( ( 1 == $count ) && ( 2 == $producer_count ) ) || ( ( $producer_count > 2 ) && ( ( $producer_count - 1 ) == $count ) ) ) {
+				$producers .= ' ' . __( 'and', 'radio-station' ) . ' ';
+			} elseif ( ( $count < $producer_count ) && ( $producer_count > 2 ) ) {
+				$producers .= ', ';
+			}
+		}
+	}
+	if ( '' != $producers ) {
+		echo '<producer>' . esc_html( $producers ) . '</producer>' . PHP_EOL;
+	}
 }