+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+= 2.5.6 =
+* Added: Filter for query and meta for show post list shortcode
+* Updated: Language translations file (.pot)
+* Updated: Bundled Dutch translation
+* Fixed: hide empty widgets in AJAX mode
+
 = 2.5.5 =
 * Updated: Freemius SDK (2.5.10)
 * Added: RSS Posts Feed: Related Show node